backend/requirements.txt
Django==3.2.19
djangorestframework==3.14.0
openpyxl==3.1.2
django-cors-headers==4.1.0
gunicorn==21.2.0
python-dotenv==1.0.0
whitenoise==6.5.0
asgiref==3.6.0
uvicorn==0.23.2